Revert "Form view"
