+++ client/store/singleProduct.js
@@ -1,8 +1,10 @@
 import axios from 'axios';
 
 const TOKEN = 'token';
+
 const GET_PRODUCT = 'GET_PRODUCT';
 const UPDATE_PRODUCT = 'UPDATE_PRODUCT';
+const DELETE_PRODUCT = 'DELETE_PRODUCT';
 
 const getProduct = (product) => {
   return {
@@ -14,6 +16,13 @@ const getProduct = (product) => {
 const updateProduct = (product) => {
   return {
     type: UPDATE_PRODUCT,
+    product,
+  };
+};
+
+const deleteProduct = (product) => {
+  return {
+    type: DELETE_PRODUCT,
     product,
   };
 };
@@ -54,6 +63,24 @@ export const fetchUpdateProduct = (product, history) => {
   };
 };
 
+export const fetchDeleteProduct = (id, history) => {
+  return async (dispatch) => {
+    try {
+      const token = window.localStorage.getItem(TOKEN);
+      if (token) {
+        const { data: product } = await axios.delete(`/api/products/${id}`, {
+          headers: {
+            authorization: token,
+          },
+        });
+        dispatch(deleteProduct(product));
+      }
+    } catch (error) {
+      console.log(error);
+    }
+  };
+};
+
 const initialState = {};
 
 export default function singleProductReducer(state = initialState, action) {
@@ -64,6 +91,9 @@ export default function singleProductReducer(state = initialState, action) {
     case UPDATE_PRODUCT:
       console.log('inside reducer update product');
       return action.product;
+    case DELETE_PRODUCT:
+      console.log('inside reducer delete product');
+      return state.filter((product) => product.id !== action.product.id);
     default:
       return state;
   }
