+++ server/api/products.js
@@ -178,10 +178,11 @@ router.post('/guest/checkout', async (req, res, next) => {
     let totalPrice = 0;
     let totalItems = 0;
     let cart = req.body.cart;
+    console.log(req.body.cart, 'cart')
     for (let i = 0; i < cart.length; i++) {
-      let product = await Product.findByPk(cart[i].product.id);
-      
-      totalItems += cart[i].quantity;
+      console.log(cart[i].price, 'ggggghghg')
+      let product = cart[i].product;
+      totalPrice += cart[i].quantity;
       totalPrice += cart[i].price;
       let productOrder = await ProductOrder.create({
         quantity: cart[i].quantity,
@@ -190,16 +191,17 @@ router.post('/guest/checkout', async (req, res, next) => {
       productOrder.setOrder(order);
       productOrder.setProduct(product);
     }
-
+console.log(totalPrice, totalItems)
     order = await order.update({
       totalPrice: totalPrice,
-      totalItems: totalItems
+      totalItems: totalItems,
     });
 
     let productOrder = await order.getProductOrders();
     for (let i = 0; i < productOrder.length; i++) {
       let id = productOrder[i].dataValues.productId;
       let product = await Product.findByPk(id);
+      console.log(productOrder.quantity, 'product');
       product = await product.update({
         inventory: product.inventory - productOrder[i].dataValues.quantity,
       });
@@ -220,7 +222,7 @@ router.put('/:id', requireToken, admin, async (req, res, next) => {
   }
 });
 
-router.delete('/:id', requireToken, async (req, res, next) => {
+router.delete('/:id', requireToken, admin, async (req, res, next) => {
   try {
     const product = await Product.findByPk(req.params.id);
     if (product) {
